requirements.txt
Pillow==10.2.0
PyMuPDF==1.23.8
pandas==2.1.4
requests==2.31.0
cloudinary==1.36.0
python-dotenv==1.0.0
